feat: add play_by_play schema
