+++ crates/nfl_play_parser/src/lib.rs
@@ -1,4 +1,4 @@
 
 pub mod error;
 pub mod schema;
-pub mod parsers;
+
