+++ crates/nfl_play_parser/src/schema/mod.rs
@@ -3,3 +3,11 @@ pub mod down_and_distance;
 pub mod game_clock;
 pub mod play_type;
 pub mod scoring_events;
+pub mod play_by_play;
+
+pub use down_and_distance::*;
+pub use game_clock::*;
+pub use play_type::*;
+pub use scoring_events::*;
+pub use play_by_play::*;
+pub use teams::*;
